feature: für die API-Kommunikation gibt es jetzt eine Funktion, die nur aufgerufen werden muss.
