+++ pp-2023/src/app/kmeans/calculateButton.js
@@ -2,10 +2,9 @@
 import {UseInputKPoints} from './input-k-points';
 import {kMeansAlgorithm} from '../utils/kmeans';
 import {HandleDynamicGeneratedInputFields} from './create-save-manuel-input';
-import {apiPostRequest, apiGetStateOfTask, apiGetResult} from './requestAPI';
+import {apiPostRequest, apiGetStateOfTask, apiGetResult, handleApiCommunication} from './requestAPI';
 import ScatterChart from './scatter-chart';
 import {returnExcel, calculateExcel} from "../utils/excelfilereader";
-import {APIError} from '../utils/userErrors';
 
 export function HandleCalculateButtonClick() {
 
@@ -46,28 +45,9 @@ export function HandleCalculateButtonClick() {
                     Der Parameter maxVersuche, gibt dabei an, wie oft ein Request wiederholt werden soll, bis dieser abbricht.
                      */
                     if (resultPost.TaskID) {
-                        try {
-                            const resultGetStateOfTask = await apiGetStateOfTask(resultPost.TaskID, 10);
-
-                            /*
-                            Liefert die apiGetStateOfTask eine 1 zurück, ist der Response erfolgreich und kann verarbeitet werden.
-                             */
-                            if (resultGetStateOfTask === 1) {
-                                const apiResult = await apiGetResult(resultPost.TaskID);
-                                // TODO handling des Resposne @Felix Kruse
-                                console.log(apiResult)
-                            }
-                        }
-                        /*
-                        Ist die Berechnung zu umfassend, dass das Zeitlimit reißt, wird dem Nutzer in dem catch-Block ein Error angezeigt.
-                         */
-                        catch (err) {
-                            const error = {
-                                "detail" : "Das Zeitlimit der Berechnung ist überschritten. " +
-                                    "Bitte verringern Sie die Anzahl an Datensätzen oder die Anzahl K."
-                            }
-                            APIError(error);
-                        }
+                        const kMeansResult = await handleApiCommunication(resultPost);
+                        //TODO Result richtig verarbeiten
+                        console.log(kMeansResult);
                     }
                 /*
                 In dem catch-Block werden allgemeine Fehler des Requests behandelt.
@@ -86,25 +66,39 @@ export function HandleCalculateButtonClick() {
                 //TODO Ladebildschirm
                 const result = await kMeansAlgorithm(inputData, kPoints);
                 ScatterChart(kPoints, chartDeletion, result);
-            } else {
-                alert('Bitte Klicken Sie auf den Button Lokal/ Serverseitig.');
             }
-            /*
-            Verarbeitung von manuell eingegeben Daten lokal.
-             */
+        /*
+        Verarbeitung von manuell eingegeben Daten lokal.
+        */
         } else if (inputDataSrc === "manuel") {
             if (localCalculation) {
-                const result = kMeansAlgorithm(inputDataArray, kPoints);
+                const result = await kMeansAlgorithm(inputDataArray, kPoints);
                 ScatterChart(kPoints, chartDeletion, result);
-                console.log(result);
                 /*
                Verarbeitung von manuell eingegeben Daten mithilfe der API.
                 */
             } else if (!localCalculation) {
-                const result = apiPostRequest(kPoints, dataArrayForWorking);
-                console.log(result);
-            } else {
-                alert('Bitte Klicken Sie auf den Button Lokal/ Serverseitig.');
+                try {
+                    /*
+                    Übersenden der eingegebenen Datei an das Backend.
+                     */
+                    const resultPost = await apiPostRequest(kPoints, false);
+
+                    /*
+                    Hier wird der Status der Task abgefragt. Aktuell wird ein Intervall von 3000 ms berücksichtigt.
+                    Der Parameter maxVersuche, gibt dabei an, wie oft ein Request wiederholt werden soll, bis dieser abbricht.
+                     */
+                    if (resultPost.TaskID) {
+                        const kMeansResult = await handleApiCommunication(resultPost);
+                        //TODO Result richtig verarbeiten
+                        console.log(kMeansResult);
+                    }
+                    /*
+                    In dem catch-Block werden allgemeine Fehler des Requests behandelt.
+                    */
+                } catch (error) {
+                    throw new error;
+                }
             }
         }
     };
